+++ polars/polars-core/src/chunked_array/ops/rolling_window.rs
@@ -93,7 +93,7 @@ mod inner_mod {
                     rolling_agg(
                         self,
                         options,
-                        rolling::sum_min_max_no_nulls::rolling_sum,
+                        rolling::no_nulls::rolling_sum,
                         rolling::nulls::rolling_sum,
                     )
                 }
@@ -106,7 +106,7 @@ mod inner_mod {
                     rolling_agg(
                         self,
                         options,
-                        rolling::sum_min_max_no_nulls::rolling_min,
+                        rolling::no_nulls::rolling_min,
                         rolling::nulls::rolling_min,
                     )
                 }
@@ -119,7 +119,7 @@ mod inner_mod {
                     rolling_agg(
                         self,
                         options,
-                        rolling::sum_min_max_no_nulls::rolling_max,
+                        rolling::no_nulls::rolling_max,
                         rolling::nulls::rolling_max,
                     )
                 }
@@ -131,8 +131,8 @@ mod inner_mod {
                     rolling_agg(
                         self,
                         options,
-                        rolling::quantile_no_nulls::rolling_median,
-                        rolling::quantile_nulls::rolling_median,
+                        rolling::no_nulls::rolling_median,
+                        rolling::nulls::rolling_median,
                     )
                 }
 
@@ -150,7 +150,7 @@ mod inner_mod {
 
                     let arr = ca.downcast_iter().next().unwrap();
                     let arr = match self.has_validity() {
-                        false => rolling::quantile_no_nulls::rolling_quantile(
+                        false => rolling::no_nulls::rolling_quantile(
                             arr.values(),
                             quantile,
                             interpolation,
@@ -159,7 +159,7 @@ mod inner_mod {
                             options.center,
                             options.weights.as_deref(),
                         ),
-                        _ => rolling::quantile_nulls::rolling_quantile(
+                        _ => rolling::nulls::rolling_quantile(
                             arr,
                             quantile,
                             interpolation,
@@ -194,7 +194,7 @@ mod inner_mod {
             rolling_agg(
                 self,
                 options,
-                rolling::sum_min_max_no_nulls::rolling_sum,
+                rolling::no_nulls::rolling_sum,
                 rolling::nulls::rolling_sum,
             )
         }
@@ -230,7 +230,7 @@ mod inner_mod {
             rolling_agg(
                 self,
                 options,
-                rolling::sum_min_max_no_nulls::rolling_min,
+                rolling::no_nulls::rolling_min,
                 rolling::nulls::rolling_min,
             )
         }
@@ -246,7 +246,7 @@ mod inner_mod {
             rolling_agg(
                 self,
                 options,
-                rolling::sum_min_max_no_nulls::rolling_max,
+                rolling::no_nulls::rolling_max,
                 rolling::nulls::rolling_max,
             )
         }
